Countdown in tracking mode.

+++ track_mode.py
@@ -34,7 +34,7 @@ from move_1_deg import *
 # Countdown in tracking mode.
 def countdown(*args):
 	print(f"Next move in {args[0]} seconds...\n")
-	os.kill(args[1], signal.SIGINT)
+	if args[0] <= 1: os.kill(args[1], signal.SIGINT)
 
 # in tracking moe,teh user searches for a body, the program goes there and then recomputes its coordonates and moves accordingly every few seconds
 # meanwhile, the user can manually fine-tune the position of the telescope so the actual position coresponds to the position the software thinks it should be pointing to
